+++ CHANGELOG.md
@@ -42,6 +42,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added CII Badge (we are compliant with the CII Best Practices)
 - Add support for doing http basic auth when loading calendars
 - Add the abilty to turn off and on the date display in the Clock Module
+- Add the abilty to set timezone on the date display in the Clock Module
 
 ### Fixed
 - Fix typo in installer.